Add files via upload
#  Airbnb-Spreadsheet-Data-analysis-project-
This project contains an analysis of the Airbnb Open Data, focusing on various trends and insights from the dataset. The analysis is performed using Excel

#TABLE OF CONTENTS:
1.Assumption & Change Log 2.Data dictionary 3.Listings 4.Calendar 5.Pivot table shows 10 most attractive neighborhood. 6.Pivot table for Most popular bedrooms for vacation rental 7.Pivot table for the number of properties most popular in harlem 8.Listing Processed data 9.Calendar Processed data 10.Pivot table show the income generated by most attractive top listing 11.Pivot table for Properties size preference for each top 10 neighborhood

# OVERVIEW
This project aims to explore, clean, and analyze the Airbnb open dataset. It includes data preprocessing, exploratory data analysis (EDA), and insights about various trends in Airbnb rentals.

# Key Questions Addressed in this Project:
1.Which neighborhoods and property sizes (i.e. number of bedrooms) are most attractive for vacation rentals?
2.How much money did these listings generate?
3.The neighborhoods which are most attractive for vacation rentals
4.The property sizes (i.e., how many bedrooms) that are most popular for vacation rentals
5.If different neighborhoods have different preferences for property sizes
#DATASET
The dataset used in this project is named NYC Airbnb Data. It contains 3 dataset such as: 1. Data dictionary 2.Listing 3.Calendar

# RESULTS
Our finding show that the best neighborhood to invest in are in Lower East side, Hells Kitchen,Harlem. People tend to rent a properties with Studio,1,2, bedrooms compared to properties that have 3 bedrooms or more. The client should invest in properties with 1-2 bedrooms in above most attractive neighborhood ie Lower East side, hells kitchen, Harlem for more profit gain. The client should focus to invest in properties with studio,1,2 bedrooms for higher profit.